ecommerce_agent/requirements.txt
google-adk==1.0.0
google-generativeai==0.8.5
google-genai
mcp==1.9.1
requests==2.32.3
beautifulsoup4==4.13.4
pillow==11.2.1
python-dotenv==1.1.0
flask==3.0.0
httpx==0.28.1